Split implementations
- Abstract caching functionality separated.
- All factory methods now belong in concretes.
- Redundant tests removed.
- Some other tests refactored.
