+++ src/AbstractBaseContainer.php
@@ -2,28 +2,14 @@
 
 namespace Dhii\Di;
 
-use Dhii\Data\Container\CreateContainerExceptionCapableTrait;
 use Dhii\Data\Container\NormalizeContainerCapableTrait;
-use Dhii\Exception\CreateInternalExceptionCapableTrait;
-use Dhii\Exception\CreateInvalidArgumentExceptionCapableTrait;
-use Dhii\Exception\CreateOutOfRangeExceptionCapableTrait;
-use Dhii\Exception\CreateRuntimeExceptionCapableTrait;
-use Dhii\Invocation\CreateInvocationExceptionCapableTrait;
-use Dhii\Invocation\CreateReflectionForCallableCapableTrait;
-use Dhii\Iterator\CountIterableCapableTrait;
-use Dhii\Iterator\ResolveIteratorCapableTrait;
 use Dhii\Util\Normalization\NormalizeArrayCapableTrait;
-use Dhii\Util\Normalization\NormalizeIntCapableTrait;
-use Dhii\Util\Normalization\NormalizeIterableCapableTrait;
 use Dhii\Util\Normalization\NormalizeStringCapableTrait;
-use Dhii\Validation\CreateValidationFailedExceptionCapableTrait;
 use Psr\Container\ContainerExceptionInterface;
 use Dhii\Util\String\StringableInterface as Stringable;
 use Exception as RootException;
 use Psr\Container\ContainerInterface as BaseContainerInterface;
 use Dhii\Data\Container\AbstractBaseContainer as BaseAbstractBaseContainer;
-use ReflectionFunction;
-use ReflectionMethod;
 
 /**
  * Common functionality for regular DI containers.
@@ -32,47 +18,11 @@ use ReflectionMethod;
  */
 abstract class AbstractBaseContainer extends BaseAbstractBaseContainer
 {
-    /* Ability to invoke callables;
-     *
-     * @since [*next-version*]
-     */
-    use InvokingTrait;
-
     /* Data object methods.
      *
      * @since [*next-version*]
      */
     use DataObjectTrait;
-
-    /* Ability to retrieve resolved cached service.
-     *
-     * @since [*next-version*]
-     */
-    use GetServiceCapableCachingTrait;
-
-    /* Awareness of a service cache.
-     *
-     * @since [*next-version*]
-     */
-    use ServiceCacheAwareTrait;
-
-    /* Ability to resolve a service definition.
-     *
-     * @since [*next-version*]
-     */
-    use ResolveDefinitionCapableTrait;
-
-    /* Ability to resolve an iterator from a Traversable chain.
-     *
-     * @since [*next-version*]
-     */
-    use ResolveIteratorCapableTrait;
-
-    /* Ability to count iterables.
-     *
-     * @since [*next-version*]
-     */
-    use CountIterableCapableTrait;
 
     /* Ability to normalize into an array.
      *
@@ -80,77 +30,17 @@ abstract class AbstractBaseContainer extends BaseAbstractBaseContainer
      */
     use NormalizeArrayCapableTrait;
 
-    /* Ability to normalize into a validator.
-     *
-     * @since [*next-version*]
-     */
-    use NormalizeIterableCapableTrait;
-
     /* Ability to normalize into a container.
      *
      * @since [*next-version*]
      */
     use NormalizeContainerCapableTrait;
 
-    /* Ability to normalize into an integer.
-     *
-     * @since [*next-version*]
-     */
-    use NormalizeIntCapableTrait;
-
     /* Ability to normalize into a string.
      *
      * @since [*next-version*]
      */
     use NormalizeStringCapableTrait;
-
-    /* Factory of Runtime exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateRuntimeExceptionCapableTrait;
-
-    /* Factory of Invalid Argument exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateInvalidArgumentExceptionCapableTrait;
-
-    /* Factory of Out of Range exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateOutOfRangeExceptionCapableTrait;
-
-    /* Factory of CachingContainer exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateContainerExceptionCapableTrait;
-
-    /* Factory of Invocation exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateInvocationExceptionCapableTrait;
-
-    /* Ability to create a reflection for a callable.
-     *
-     * @since [*next-version*]
-     */
-    use CreateReflectionForCallableCapableTrait;
-
-    /* Factory of Validation Failed exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateValidationFailedExceptionCapableTrait;
-
-    /* Factory of Internal exception.
-     *
-     * @since [*next-version*]
-     */
-    use CreateInternalExceptionCapableTrait;
 
     /**
      * {@inheritdoc}
@@ -182,22 +72,16 @@ abstract class AbstractBaseContainer extends BaseAbstractBaseContainer
     }
 
     /**
-     * {@inheritdoc}
+     * Retrieves a service by key.
      *
      * @since [*next-version*]
-     */
-    protected function _createReflectionFunction($functionName)
-    {
-        return new ReflectionFunction($functionName);
-    }
-
-    /**
-     * {@inheritdoc}
      *
-     * @since [*next-version*]
+     * @param string|int|float|bool|Stringable $key The key, for which to get the service.
+     *
+     * @throw NotFoundExceptionInterface If no service or definition found for key.
+     * @throw ContainerExceptionInterface If service or service definition could not be retrieved.
+     *
+     * @return mixed The corresponding service.
      */
-    protected function _createReflectionMethod($className, $methodName)
-    {
-        return new ReflectionMethod($className, $methodName);
-    }
+    abstract protected function _getService($key);
 }
