+++ src/CachingContainer.php
@@ -3,10 +3,18 @@
 namespace Dhii\Di;
 
 use ArrayAccess;
-use Dhii\Cache\ContainerInterface as CacheContainerInterface;
-use Dhii\I18n\StringTranslatingTrait;
 use stdClass;
 use Psr\Container\ContainerInterface as BaseContainerInterface;
+use Dhii\Data\Container\CreateContainerExceptionCapableTrait;
+use Dhii\Exception\CreateInvalidArgumentExceptionCapableTrait;
+use Dhii\Exception\CreateOutOfRangeExceptionCapableTrait;
+use Dhii\Exception\CreateRuntimeExceptionCapableTrait;
+use Dhii\Exception\CreateInternalExceptionCapableTrait;
+use Dhii\Invocation\CreateInvocationExceptionCapableTrait;
+use Dhii\Validation\CreateValidationFailedExceptionCapableTrait;
+use Dhii\I18n\StringTranslatingTrait;
+use Dhii\Cache\ContainerInterface as CacheContainerInterface;
+use Dhii\Util\Normalization\NormalizeIntCapableTrait;
 
 /**
  * A basic DI container.
@@ -15,7 +23,7 @@ use Psr\Container\ContainerInterface as BaseContainerInterface;
  *
  * @since [*next-version*]
  */
-class CachingContainer extends AbstractBaseContainer
+class CachingContainer extends AbstractBaseCachingContainer
 {
     /*
      * Basic ability to i18n strings.
@@ -23,6 +31,54 @@ class CachingContainer extends AbstractBaseContainer
      * @since [*next-version*]
      */
     use StringTranslatingTrait;
+
+    /* Ability to normalize into an integer.
+     *
+     * @since [*next-version*]
+     */
+    use NormalizeIntCapableTrait;
+
+    /* Factory of Validation Failed exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateValidationFailedExceptionCapableTrait;
+
+    /* Factory of Runtime exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateRuntimeExceptionCapableTrait;
+
+    /* Factory of Invalid Argument exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateInvalidArgumentExceptionCapableTrait;
+
+    /* Factory of Out of Range exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateOutOfRangeExceptionCapableTrait;
+
+    /* Factory of Invocation exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateInvocationExceptionCapableTrait;
+
+    /* Factory of Internal exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateInternalExceptionCapableTrait;
+
+    /* Factory of CachingContainer exception.
+     *
+     * @since [*next-version*]
+     */
+    use CreateContainerExceptionCapableTrait;
 
     /**
      * @since [*next-version*]
@@ -38,15 +94,5 @@ class CachingContainer extends AbstractBaseContainer
 
         $this->_setDataStore($services);
         $this->_setServiceCache($serviceCache);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @since [*next-version*]
-     */
-    protected function _getArgsForDefinition($definition)
-    {
-        return [$this];
     }
 }
